Removed tkinter gui.
Tkinter is not running on the target for unknown reasons.
We wil not need the script theremost probably but it should work anyway.
The gui version will still be available in tkinter branch.

We will have to rebase it here and then, because actuale development will happen on master branch.

Signed-off-by: bhamacher <b.hamacher@zera.de>